remove context_to_jones_array - move jones_baseline_view_to_imageset and flag_baseline_view_to_flagmask into flags - separate allocation and initialisation of flags and jones - don't use FlagContext if flags are blank - no more flag_ctx.to_array()
