+++ benches/bench.rs
@@ -1,5 +1,4 @@
 use birli::{
-    context_to_jones_array,
     flags::{add_dimension, flag_to_weight_array, get_weight_factor},
     io::write_uvfits,
     write_ms, VisSelection,
@@ -49,14 +48,15 @@ fn bench_uvfits_output_1196175296_none(crt: &mut Criterion) {
     let mut vis_sel = VisSelection::from_mwalib(&corr_ctx).unwrap();
 
     vis_sel.timestep_range = 0..NUM_TIMESTEPS;
-    let (jones_array, flag_array) = context_to_jones_array(
-        &corr_ctx,
-        &vis_sel.timestep_range,
-        &vis_sel.coarse_chan_range,
-        None,
-        false,
-    )
-    .unwrap();
+
+    let fine_chans_per_coarse = corr_ctx.metafits_context.num_corr_fine_chans_per_coarse;
+    let mut flag_array = vis_sel.allocate_flags(fine_chans_per_coarse).unwrap();
+    let mut jones_array = vis_sel
+        .allocate_jones(corr_ctx.metafits_context.num_corr_fine_chans_per_coarse)
+        .unwrap();
+    vis_sel
+        .read_mwalib(&corr_ctx, &mut jones_array, &mut flag_array, false)
+        .unwrap();
 
     let weight_factor = get_weight_factor(&corr_ctx);
     let flag_array = add_dimension(flag_array.view(), 4);
@@ -95,14 +95,14 @@ fn bench_ms_output_1196175296_none(crt: &mut Criterion) {
     let mut vis_sel = VisSelection::from_mwalib(&corr_ctx).unwrap();
 
     vis_sel.timestep_range = 0..NUM_TIMESTEPS;
-    let (jones_array, flag_array) = context_to_jones_array(
-        &corr_ctx,
-        &vis_sel.timestep_range,
-        &vis_sel.coarse_chan_range,
-        None,
-        false,
-    )
-    .unwrap();
+    let fine_chans_per_coarse = corr_ctx.metafits_context.num_corr_fine_chans_per_coarse;
+    let mut flag_array = vis_sel.allocate_flags(fine_chans_per_coarse).unwrap();
+    let mut jones_array = vis_sel
+        .allocate_jones(corr_ctx.metafits_context.num_corr_fine_chans_per_coarse)
+        .unwrap();
+    vis_sel
+        .read_mwalib(&corr_ctx, &mut jones_array, &mut flag_array, false)
+        .unwrap();
 
     let weight_factor = get_weight_factor(&corr_ctx);
     let flag_array = add_dimension(flag_array.view(), 4);
