+++ src/preprocessing.rs
@@ -242,7 +242,7 @@ mod tests {
     use tempfile::tempdir;
 
     use crate::{
-        add_dimension, context_to_jones_array, flag_to_weight_array,
+        add_dimension, flag_to_weight_array,
         flags::get_weight_factor,
         test_common::{compare_uvfits_with_csv, get_1254670392_avg_paths},
         write_uvfits, FlagContext, VisSelection,
@@ -279,23 +279,20 @@ mod tests {
         let (avg_time, avg_freq) = (1, 1);
 
         let flag_ctx = FlagContext::from_mwalib(&corr_ctx);
-
-        let flag_array = flag_ctx
-            .to_array(
+        let fine_chans_per_coarse = corr_ctx.metafits_context.num_corr_fine_chans_per_coarse;
+        let mut flag_array = vis_sel.allocate_flags(fine_chans_per_coarse).unwrap();
+        flag_ctx
+            .set_flags(
+                &mut flag_array,
                 &vis_sel.timestep_range,
                 &vis_sel.coarse_chan_range,
                 vis_sel.get_ant_pairs(&corr_ctx.metafits_context),
             )
             .unwrap();
-
-        let (mut jones_array, mut flag_array) = context_to_jones_array(
-            &corr_ctx,
-            &vis_sel.timestep_range,
-            &vis_sel.coarse_chan_range,
-            Some(flag_array),
-            prep_ctx.draw_progress,
-        )
-        .unwrap();
+        let mut jones_array = vis_sel.allocate_jones(fine_chans_per_coarse).unwrap();
+        vis_sel
+            .read_mwalib(&corr_ctx, &mut jones_array, &mut flag_array, false)
+            .unwrap();
 
         // generate weights
         let weight_factor = get_weight_factor(&corr_ctx);
